Allow setting the bounds of decryt
Also add '-in' for specifying file to decrypt (or encrypt).

File input was needed to use true random access.

If range is specifed for file, only the salt in the first block and the range are read.
For stdin, the other data is skipped or not used.

In both cases, the AES-CTR counter is updated accordingly.